+++ 01/p5.c
@@ -24,13 +24,13 @@ int main(int argc, char **argv){
 
   for(i=0; i<n; i++)
     for(j=0; j<n; j++)
-      printf(j <n-1 ? "%d ":"%d\n", (int)nnmat[0][i*n+j]);
+      printf(j <n-1 ? "%.1f ":"%.1f\n", nnmat[0][i*n+j]);
 
   printf("\nMatrix with 2d indexing:\n");
 
   for(i=0; i<n; i++)
     for(j=0; j<n; j++)
-      printf(j <n-1 ? "%d ":"%d\n", (int)nnmat[i][j]);
+      printf(j <n-1 ? "%.1f ":"%.1f\n", nnmat[i][j]);
 
   return EXIT_SUCCESS;
 }
